pyproject.toml
[project]
name = "application-pipeline"
version = "0.1.0"
description = "Career application pipeline — structured conversion tracking, narrative composition, and analytics"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "pyyaml",
    "ruamel.yaml",
    "typer",
    "google-genai",
    "mcp",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "ruff",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["scripts.cli", "scripts.standup", "scripts.pipeline_status", "scripts.score", "scripts.compose"]

[project.scripts]
pipeline = "scripts.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "synthetic: tests that perform live API calls to ATS platforms"
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by line-length)
    "E741",  # ambiguous variable name
]

[tool.ruff.lint.per-file-ignores]
"scripts/legacy-submission/*" = ["E", "F"]
"tests/*" = ["E402"]
